<refactor> 중복파일 정리
